different approach for trimesh compat stuff
