+++ docs/reference.rst
@@ -38,7 +38,7 @@ Currently this includes the following classes, which comprise the public API:
     ~pygfx.geometries.Geometry
     ~pygfx.geometries.TextGeometry
     ~pygfx.geometries.TextItem
-    ~pygfx.geometries.trimesh_geometry
+    ~pygfx.geometries.geometry_from_trimesh
     ~pygfx.geometries.octahedron_geometry
     ~pygfx.geometries.icosahedron_geometry
     ~pygfx.geometries.dodecahedron_geometry
@@ -54,7 +54,7 @@ Currently this includes the following classes, which comprise the public API:
     ~pygfx.helpers.SpotLightHelper
 
     ~pygfx.materials.Material
-    ~pygfx.materials.trimesh_material
+    ~pygfx.materials.material_from_trimesh
     ~pygfx.materials.MeshBasicMaterial
     ~pygfx.materials.MeshPhongMaterial
     ~pygfx.materials.MeshNormalMaterial
