+++ pygfx/materials/__init__.py
@@ -14,7 +14,7 @@ colormaps, the strength of specular reflections, etc.
     :template: ../_templates/custom_layout.rst
 
     Material
-    trimesh_material
+    material_from_trimesh
 
     MeshBasicMaterial
     MeshPhongMaterial
@@ -51,7 +51,7 @@ colormaps, the strength of specular reflections, etc.
 # flake8: noqa
 
 from ._base import Material
-from ._compat import trimesh_material
+from ._compat import material_from_trimesh
 from ._mesh import (
     MeshBasicMaterial,
     MeshPhongMaterial,
@@ -92,4 +92,4 @@ __all__ = [
 __all__.sort()
 __all__.remove("Material")
 __all__.insert(0, "Material")
-__all__.extend(["trimesh_material"])
+__all__.extend(["material_from_trimesh"])
